+++ NEWS.md
@@ -1,3 +1,5 @@
+# mlr3misc (development version)
+
 # mlr3misc 0.15.1
 
 * refactor: Update `as_callback()` functions.