+++ NEWS.md
@@ -1,6 +1,7 @@
 # mlr3misc (development version)
 
 * feat: `as_callbacks()` returns a list named by the callback ids now.
+* feat: Added logical operators `%check&&%` and `%check||%` for `check_*`-functions from `checkmate` (moved here from `mlr3pipelines`).
 
 # mlr3misc 0.16.0