+++ NEWS.md
@@ -4,6 +4,7 @@
 * feat: Added logical operators `%check&&%` and `%check||%` for `check_*`-functions from `checkmate` (moved here from `mlr3pipelines`).
 * Added helper `cat_cli()`.
 * fix: default `dicts_suggest = NULL` for `dictionary_get_item()` and `dictionary_retrieve_item()` for backward compatibility.
+* fix a bug in `dictionary_sugar_inc_get`
 
 # mlr3misc 0.16.0